+++ test/template_test_helpers.rb
@@ -1,12 +1,17 @@
 module TemplateTestHelpers
   def around(&block)
-    Dir.mktmpdir do |temporary_directory|
-      @partial_path = Pathname(temporary_directory).join("app", "views")
-      @config_path = Pathname(temporary_directory).join("config")
+    Dir.mktmpdir do |tmpdir|
+      temporary_directory = Pathname(tmpdir)
+
+      @partial_path = temporary_directory.join("app", "views")
+      @config_path = temporary_directory.join("config")
+      @fixture_path = temporary_directory.join("test", "clients", "fixtures")
 
       with_view_path_prefixes(@partial_path) do
         with_config_path(@config_path) do
-          block.call
+          with_fixture_path_prefixes(@fixture_path) do
+            block.call
+          end
         end
       end
     end
@@ -23,25 +28,37 @@ module TemplateTestHelpers
   end
 
   def with_view_path_prefixes(temporary_view_directory, &block)
-    view_paths = ActionClient::Base.view_paths
+    with_path_prefixes(ActionClient::Base, temporary_view_directory, &block)
+  end
 
-    ActionClient::Base.prepend_view_path(temporary_view_directory)
+  def with_fixture_path_prefixes(temporary_view_directory, &block)
+    with_path_prefixes(ActionClient::Test::Client, temporary_view_directory, &block)
+  end
+
+  def with_path_prefixes(client_class, temporary_view_directory, &block)
+    view_paths = client_class.view_paths
+
+    client_class.prepend_view_path(temporary_view_directory)
 
     block.call
   ensure
-    ActionClient::Base.view_paths = view_paths
+    client_class.view_paths = view_paths
   end
 
-  def declare_config(partial_path, body)
-    @config_path.join(partial_path).tap do |file|
-      file.dirname.mkpath
-
-      file.write(body)
-    end
+  def declare_fixture(path, body)
+    declare_file(@fixture_path, path, body)
   end
 
-  def declare_template(partial_path, body)
-    @partial_path.join(partial_path).tap do |file|
+  def declare_template(path, body)
+    declare_file(@partial_path, path, body)
+  end
+
+  def declare_config(path, body)
+    declare_file(@config_path, path, body)
+  end
+
+  def declare_file(directory, path, body)
+    directory.join(path).tap do |file|
       file.dirname.mkpath
 
       file.write(body)